Refactor collapsable component and add defaultClosed prop
Refactor the collapsable component in the src/components/collapsable.tsx file to include a new prop called defaultClosed. This prop allows the component to be initially closed when rendered. The default value is set to false if not provided. This change improves the flexibility and usability of the collapsable component.

Update collapsable usage in search page

In the src/routes/search/[query]/index.tsx file, update the usage of the collapsable component for different sections (tdk, nisanyan, kubbealti, luggat, benzer, and rhyme). Each section now includes a defaultClosed prop that determines whether the section should be initially closed or not. This change enhances the user experience by providing a more intuitive interface for collapsing and expanding sections.